Changes as per prior code review
