+++ src/test/java/com/microsoft/sqlserver/testframework/Constants.java
@@ -180,6 +180,10 @@ public final class Constants {
     public static final String PREPARE_METHOD = "PREPAREMETHOD";
     public static final String CONFIG_PROPERTIES_FILE = "config.properties";
     public static final String UTF8 = "UTF-8";
+    public static final double MAX_VALUE_MONEY = 922337203685477.5807;
+    public static final double MIN_VALUE_MONEY = -922337203685477.5808;
+    public static final double MAX_VALUE_SMALLMONEY = 214748.3647;
+    public static final double MIN_VALUE_SMALLMONEY = -214748.3648;
 
     public enum LOB {
         CLOB,
